Move worker script
