+++ src/app/core/worker/functions.py
@@ -1,17 +1,11 @@
 import asyncio
 import logging
 import uvloop
-from arq.connections import RedisSettings
 from arq.worker import Worker
-
-from .core.config import settings
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-
-REDIS_QUEUE_HOST = settings.REDIS_QUEUE_HOST
-REDIS_QUEUE_PORT = settings.REDIS_QUEUE_PORT
 
 
 # -------- background tasks --------
@@ -27,12 +21,3 @@ async def startup(ctx: Worker) -> None:
 
 async def shutdown(ctx: Worker) -> None:
     logging.info("Worker end")
-
-
-# -------- class --------
-class WorkerSettings:
-    functions = [sample_background_task]
-    redis_settings = RedisSettings(host=REDIS_QUEUE_HOST, port=REDIS_QUEUE_PORT)
-    on_startup = startup
-    on_shutdown = shutdown
-    handle_signals = False
