+++ src/app/core/worker/settings.py
@@ -1,0 +1,14 @@
+from arq.connections import RedisSettings
+from ...core.config import settings
+from .functions import sample_background_task, startup, shutdown
+
+REDIS_QUEUE_HOST = settings.REDIS_QUEUE_HOST
+REDIS_QUEUE_PORT = settings.REDIS_QUEUE_PORT
+
+
+class WorkerSettings:
+    functions = [sample_background_task]
+    redis_settings = RedisSettings(host=REDIS_QUEUE_HOST, port=REDIS_QUEUE_PORT)
+    on_startup = startup
+    on_shutdown = shutdown
+    handle_signals = False
